Many modifications. Changes were primarily geared towards being able to insert new dives, edit them, and view a listing of dives for the current user. Some database functions were added/modified. there were changes to common.php and session.php to facilitate this. In particular, session.php was modified so that two new functions were added to control authentication. getAuthStatus() is now used to check if the user is logged in and allowed to do "stuff". This did require making a call to the database on each hit (other CM systems do this too, so maybe it's not such a big deal?). Also of note, the CSS file was fixed so things may look a bit different now.
git-svn-id: svn://chernobyl.imartin.net/divelog/trunk@23 51321470-acf5-0310-bcb5-c959670cf2a8

+++ site/logout.php
@@ -4,7 +4,7 @@ include('session.php');
 global $session;
 
 $uid = $session->getID();
-if( !isset($uid) || $uid == "" ){
+if( isset($uid) && $uid != "" ){
 	$session->destroy($uid);
 }
 header("Location: index.php");
